Delete some unused file and fix the structure
