--- recognition/ResNet.py
@@ -1,78 +1,0 @@
-import torch
-import torch.nn as nn
-
-
-class ResNet(nn.Module):
-    def __init__(self, block, num_block):
-        super(ResNet, self).__init__()
-        self.in_channels = 64
-
-        self.conv1 = nn.Sequential(
-            nn.Conv2d(3, 64, kernel_size=7, stride=2, padding=3, bias=False),
-            nn.BatchNorm2d(64),
-            nn.ReLU(inplace=True)
-        )
-        self.maxpool = nn.MaxPool2d(kernel_size=3, stride=2, padding=1)
-        self.conv2_x = self._make_layer(block, 64, num_block[0], 1)
-        self.conv3_x = self._make_layer(block, 128, num_block[1], 2)
-        self.conv4_x = self._make_layer(block, 256, num_block[2], 2)
-        self.conv5_x = self._make_layer(block, 512, num_block[3], 2)
-
-    def _make_layer(self, block, out_channels, num_blocks, stride):
-        """
-
-        :param block:  block type, basic block or bottle_neck block
-        :param out_channels:  output channels of each block
-        :param num_blocks:  how many blocks per layer
-        :param stride:  stride
-        :return:  return a resnet layer
-        """
-        strides = [stride] + [1] * (num_blocks - 1)
-        layers = []
-        for stride in strides:
-            layers.append(block(self.in_channels, out_channels, stride))
-            self.in_channels = out_channels * block.expansion
-        return nn.Sequential(*layers)
-
-    def forward(self, x):
-        output = self.conv1(x)
-        output = self.maxpool(output)
-        output = self.conv2_x(output)
-        output = self.conv3_x(output)
-        output = self.conv4_x(output)
-        output = self.conv5_x(output)
-        return output
-
-
-class Bottleneck(nn.Module):
-    expansion = 4
-
-    def __init__(self, in_channels, out_channels, stride=1):
-        super(Bottleneck, self).__init__()
-        self.residual_function = nn.Sequential(
-            nn.Conv2d(in_channels, out_channels, kernel_size=1, bias=False),
-            nn.BatchNorm2d(out_channels),
-            nn.ReLU(inplace=True),
-            nn.Conv2d(out_channels, out_channels, stride=stride, kernel_size=3, padding=1, bias=False),
-            nn.BatchNorm2d(out_channels),
-            nn.ReLU(inplace=True),
-            nn.Conv2d(out_channels, out_channels * Bottleneck.expansion, kernel_size=1, bias=False),
-            nn.BatchNorm2d(out_channels * Bottleneck.expansion)
-        )
-
-        self.shortcut = nn.Sequential()
-        if stride != 1 or in_channels != out_channels * Bottleneck.expansion:
-            self.shortcut = nn.Sequential(
-                nn.Conv2d(in_channels, out_channels * Bottleneck.expansion, stride=stride, kernel_size=1, bias=False),
-                nn.BatchNorm2d(out_channels * Bottleneck.expansion)
-            )
-
-    def forward(self, x):
-        return nn.ReLU(inplace=True)(self.residual_function(x) + self.shortcut(x))
-
-
-# test
-# if __name__ == '__main__':
-#     net = ResNet(Bottleneck, [3, 4, 6, 3])
-#     y = net(torch.randn(1, 3, 256, 256))
-#     print(y.size())  # 结果应该是 torch.Size([1, 2048, 8, 8])
