+++ recognition/train.py
@@ -1,42 +1,118 @@
-# train.py
+"""
+    File name: modules.py
+    Author: Fanhao Zeng
+    Date created: 11/10/2023
+    Date last modified: 16/10/2023
+    Python Version: 3.10.12
+"""
+
+import os
+from torch.utils.tensorboard import SummaryWriter
 import torch
-from torch.utils.data import DataLoader
-from dataset import ISICDataset  # 确保这一行正确地导入了您的数据集类
-from modules import MaskRCNNModule
+import torch.optim as optim
+from torch import nn
+from torch.utils.data import DataLoader, Subset
+from modules import SiameseNetwork
+from dataset import get_train_dataset
 
-# 初始化模型
-mask_rcnn_model = MaskRCNNModule(pretrained=False)
-model = mask_rcnn_model.get_model()
-model.train()
 
-# 创建数据加载器
-train_dataset = ISICDataset(path="E:/comp3710/ISIC2018", type="Training")
-train_loader = DataLoader(train_dataset, batch_size=4, shuffle=True)
+def train(model, dataloader, device, optimizer, epoch):
 
-# 定义优化器和损失函数（如果需要的话）
-optimizer = torch.optim.SGD(model.parameters(), lr=0.001, momentum=0.9)
+    model.train()
 
-# 训练循环
-for epoch in range(10):  # 假设我们训练10个epoch
-    for images, targets in train_loader:
-        # 模型预测
-        loss_dict = model(images, targets)
+    # Using `BCELoss` as the loss function
+    criterion = nn.BCELoss()
 
-        # 计算总损失
-        losses = sum(loss for loss in loss_dict.values())
+    correct = 0  # Reset correct count for each epoch
+    total_samples = 0  # Keep track of total samples processed
 
-        # 清除之前的梯度
-        optimizer.zero_grad()
-
-        # 反向传播
-        losses.backward()
-
-        # 更新权重
+    for batch_idx, (images_1, images_2, targets) in enumerate(dataloader):
+        images_1, images_2, targets = images_1.to(device), images_2.to(device), targets.to(device)
+        optimizer.zero_grad() # Zero out gradients
+        outputs = model(images_1, images_2).squeeze()
+        loss = criterion(outputs, targets) # Calculate loss
+        loss.backward()
         optimizer.step()
 
-        print(f"Loss: {losses.item()}")
+        # Calculate accuracy
+        pred = torch.where(outputs > 0.5, torch.tensor(1.0, device=device), torch.tensor(0.0, device=device))
+        correct += pred.eq(targets.view_as(pred)).sum().item()
+        total_samples += len(targets)  # Update total samples
 
-    print(f"Epoch {epoch + 1} completed")
+        if batch_idx % 10 == 0:
+            accuracy = 100. * correct / total_samples  # Calculate accuracy
 
-# 保存模型
-torch.save(model.state_dict(), "mask_rcnn_model.pth")
+            print('Train Epoch: {} [{}/{} ({:.0f}%)]\tLoss: {:.6f}\tAccuracy: {:.2f}%'.format(
+                epoch, total_samples, len(dataloader.dataset),
+                100. * batch_idx / len(dataloader), loss.item(), accuracy))
+
+            # Log to TensorBoard
+            writer.add_scalar('Training Loss', loss.item(), epoch * len(dataloader) + batch_idx)
+            writer.add_scalar('Training Accuracy', accuracy, epoch * len(dataloader) + batch_idx)
+
+
+def validate(model, dataloader, device):
+    print("Starting Validation.")
+    model.eval()
+    test_loss = 0
+    correct = 0
+
+    criterion = nn.BCELoss()
+
+    with torch.no_grad():
+        for (images_1, images_2, targets) in dataloader:
+            images_1, images_2, targets = images_1.to(device), images_2.to(device), targets.to(device)
+            outputs = model(images_1, images_2).squeeze()
+            test_loss += criterion(outputs, targets).sum().item()  # sum up batch loss
+            pred = torch.where(outputs > 0.5, 1, 0)  # get the index of the max log-probability
+            correct += pred.eq(targets.view_as(pred)).sum().item()
+
+    test_loss /= len(dataloader.dataset)
+
+    print('\nVal set: Average loss: {:.4f}, Validation Accuracy: {}/{} ({:.0f}%)\n'.format(
+        test_loss, correct, len(dataloader.dataset),
+        100. * correct / len(dataloader.dataset)))
+
+    print("Finished Validation.")
+
+
+if __name__ == '__main__':
+    # Initialize TensorBoard
+    writer = SummaryWriter('logs/siamese_experiment')
+
+    epochs = 50
+    batch_size = 256
+    learning_rate = 0.001
+
+    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+    print(device)
+
+    # Initialize model, criterion, and optimizer
+    model = SiameseNetwork().to(device)
+    optimizer = optim.Adam(model.parameters(), lr=learning_rate)
+
+    # Get dataloaders
+    print("Loading data...")
+    train_data = get_train_dataset('E:/comp3710/AD_NC')
+    train_dataloader = torch.utils.data.DataLoader(train_data, batch_size=batch_size)
+    print("Data loaded.")
+
+    save_directory = "E:/PatternAnalysis-2023/results"
+    save_filename = f"siamese_network_{epochs}epochs.pt"
+
+    print("Starting training.")
+    for epoch in range(1, epochs + 1):
+        train(model, train_dataloader, device, optimizer, epoch)
+        # validate(model, val_loader, device)
+    print("Finished training.")
+
+    # create directory if it doesn't exist
+    if not os.path.exists(save_directory):
+        os.makedirs(save_directory)
+
+    # save model
+    save_path = os.path.join(save_directory, save_filename)
+    torch.save(model.state_dict(), save_path)
+
+    writer.close()
+
